Rearrange code for easier jest testing
