+++ elisa/src/client/RunDetails/utils.ts
@@ -1,8 +1,16 @@
 import { Utils, getServerContext, ActionURL } from '@labkey/api';
 import { naturalSort } from "@labkey/components";
 
-import { CONTROL_COL_NAME, ID_COL_NAME, SAMPLE_COL_NAME, STANDARDS_LABEL, WELL_GROUP_COL_NAME } from "./constants";
-import { PlotOptions, SelectOptions } from "./models";
+import {
+    CONTROL_COL_NAME,
+    HOVER_COLUMN_NAMES,
+    ID_COL_NAME, PLOT_HEIGHT,
+    SAMPLE_COL_NAME,
+    STANDARDS_LABEL,
+    WELL_GROUP_COL_NAME,
+    X_AXIS_PROP, Y_AXIS_PROP
+} from "./constants";
+import { CurveFitData, PlotOptions, SelectOptions } from "./models";
 
 const valExponentialDigits = 6;
 export function tickFormatFn(value) {
@@ -10,13 +18,6 @@ export function tickFormatFn(value) {
         return value.toExponential();
     }
     return value;
-}
-
-export function exportSVGToFile(svgEl, format, title: string) {
-    if (svgEl && svgEl.length > 0) {
-        const LABKEY = getServerContext();
-        LABKEY.vis.SVGConverter.convert(svgEl[0], format, 'Calibration Curve - ' + title);
-    }
 }
 
 export function filterDataByPlotOptions(data: any[], samples: string[], controls: string[], options: PlotOptions, includeIdFilter = true): any[] {
@@ -32,13 +33,19 @@ export function filterDataByPlotOptions(data: any[], samples: string[], controls
     });
 }
 
-export function getMinFromData(data: any[], prop: string): number {
-    const values = data.map((row) => row[prop] as number);
+export function getMinFromData(data: any[], prop: string, includeNull = true): number {
+    let values = data.map((row) => row[prop] as number);
+    if (!includeNull) {
+        values = values.filter(d => d !== null);
+    }
     return Math.min(...values);
 }
 
-export function getMaxFromData(data: any[], prop: string): number {
-    const values = data.map((row) => row[prop] as number);
+export function getMaxFromData(data: any[], prop: string, includeNull = true): number {
+    let values = data.map((row) => row[prop] as number);
+    if (!includeNull) {
+        values = values.filter(d => d !== null);
+    }
     return Math.max(...values);
 }
 
@@ -76,10 +83,10 @@ export function getResultsViewURL(protocolId: number, runId: number, plotOptions
         // add in an extra separator to make this an OR between the samples and controls
         if (controls.indexOf(STANDARDS_LABEL) > -1) {
             if (showAllSamples || samples.length > 0) {
-                params['Data.' + WELL_GROUP_COL_NAME + '~containsoneof'] = STANDARDS_LABEL + ';Sample';
+                params['Data.' + WELL_GROUP_COL_NAME + '~containsoneof'] = 'Standards;Sample';
             }
             else {
-                params['Data.' + WELL_GROUP_COL_NAME + '~in'] = STANDARDS_LABEL;
+                params['Data.' + WELL_GROUP_COL_NAME + '~eq'] = 'Standards';
             }
         }
         else {
@@ -104,4 +111,119 @@ export function getUniqueIdsForPlotSelections(data: any[], plotOptions: PlotOpti
 
 export function shouldReloadCurveFitData(prevPlotOptions: PlotOptions, newPlotOptions: PlotOptions): boolean {
     return prevPlotOptions.plateName !== newPlotOptions.plateName || prevPlotOptions.spot !== newPlotOptions.spot;
+}
+
+export function getPlotConfigFromOptions(
+    renderId: string, width: number, plotOptions: PlotOptions, data: any[], curveFitData: CurveFitData,
+    runName: string, columnInfo: {[key: string]: any}
+): {[key: string]: any} {
+    const LABKEY = getServerContext();
+
+    const aes = {
+        x: X_AXIS_PROP,
+        y: Y_AXIS_PROP,
+        hoverText: function(row){
+            return HOVER_COLUMN_NAMES.map((col) => {
+                return col + ': ' + row[col];
+            }).join('\n');
+        }
+    };
+    if (plotOptions.showLegend) {
+        aes['color'] = ID_COL_NAME;
+    }
+
+    const layers = [];
+    if (data !== undefined) {
+        layers.push(
+            new LABKEY.vis.Layer({
+                data,
+                geom: new LABKEY.vis.Geom.Point({
+                    plotNullPoints: true,
+                    size: 5,
+                    opacity: .5,
+                    color: '#116596'
+                }),
+                aes
+            })
+        );
+    }
+    if (plotOptions.showCurve && curveFitData) {
+        layers.push(
+            new LABKEY.vis.Layer({
+                geom: new LABKEY.vis.Geom.Path({
+                    size: 2,
+                    color: '#555'
+                }),
+                aes: {x: 'x', y: 'y'},
+                data: curveFitData.points
+            })
+        );
+    }
+
+    const margins = {};
+    if (!plotOptions.showLegend) {
+        margins['right'] = 10;
+    }
+
+    const scales = {
+        x: {scaleType: 'continuous', trans: plotOptions.xAxisScale, tickFormat: tickFormatFn},
+        y: {scaleType: 'continuous', trans: plotOptions.yAxisScale, tickFormat: tickFormatFn}
+    };
+
+    return {
+        renderTo: renderId,
+        width,
+        height: PLOT_HEIGHT,
+        layers,
+        labels: {
+            main: { value: getPlotTitle(runName, plotOptions) },
+            x: { value: columnInfo[X_AXIS_PROP] ? columnInfo[X_AXIS_PROP].caption : X_AXIS_PROP },
+            y: { value: columnInfo[Y_AXIS_PROP] ? columnInfo[Y_AXIS_PROP].caption : Y_AXIS_PROP }
+        },
+        margins,
+        scales
+    };
+}
+
+export function getColumnInfoFromQueryDetails(queryInfo: {[key: string]: any}): {[key: string]: any} {
+    const columnInfo = {};
+    queryInfo.columns.forEach((col) => {
+        columnInfo[col.fieldKey] = col;
+    });
+
+    return columnInfo;
+}
+
+export function parsePlotDataFromResponse(response: {[key: string]: any}): any[] {
+    const data = [];
+    response.rows.forEach((row) => {
+        // consolidate the control name for the standard wells so they plot the same color
+        const isStandard = row[WELL_GROUP_COL_NAME] === 'Standards';
+        if (isStandard) {
+            row[CONTROL_COL_NAME] = STANDARDS_LABEL;
+        }
+
+        // if the row has a control value, make sure the sample column is null
+        if (row[CONTROL_COL_NAME]) {
+            row[SAMPLE_COL_NAME] = null;
+        }
+
+        row[ID_COL_NAME] = row[CONTROL_COL_NAME] || row[SAMPLE_COL_NAME];
+        data.push(row);
+    });
+
+    return data;
+}
+
+export function getUpdatedPlotOptions(key: string, value: any, resetIdSelection: boolean, plotOptions: PlotOptions): PlotOptions {
+    const updatedPlotOptions = {
+        ...plotOptions,
+        [key]: value
+    };
+
+    updatedPlotOptions.showAllSamples = resetIdSelection ? true : updatedPlotOptions.showAllSamples;
+    updatedPlotOptions.samples = resetIdSelection || updatedPlotOptions.showAllSamples ? [] : updatedPlotOptions.samples;
+    updatedPlotOptions.showAllControls = resetIdSelection ? true : updatedPlotOptions.showAllControls;
+    updatedPlotOptions.controls = resetIdSelection || updatedPlotOptions.showAllControls ? [] : updatedPlotOptions.controls;
+    return updatedPlotOptions;
 }
